requirements/base.txt
Django==1.6.4
South==0.8.4
dj-database-url==0.3.0
psycopg2==2.5.2
docutils==0.11
django-admin-bootstrapped==1.6.4
djangorestframework==2.3.13
static==0.4
wsgiref==0.1.2